split out rakefile_transform
The task that generates the ep-popolo file.
